create examples directory & module; move doobie example into examples;
