+++ docs/logging.md
@@ -73,7 +73,7 @@ functionality:
 * `ToStringLoggable` for using `.toString` for logging
 * `HideLoggable` to exclude value from logging
 
-### Loggable example
+### Loggable tofu.example
 
 ```scala:reset
 import tofu.logging._
@@ -90,7 +90,7 @@ implicit val userLoggable = new DictLoggable[User] {
 }
 ```
 
-Let's take a look at this example.  
+Let's take a look at this tofu.example.  
 First, we define a loggable for our `User` class as a `DictLoggable` which means that we want to log it as a multi-field
 object with structure.  
 Second, we define two methods that describe how `User` should be logged:
@@ -181,7 +181,7 @@ exist though).
 ### Plain `Logging`
 
 * by asking that name from user:
-  For example here we create named instance and pass it explicitly.
+  For tofu.example here we create named instance and pass it explicitly.
 
 ```scala:reset
 import tofu.logging._
@@ -435,7 +435,7 @@ Moreover `Server` itself logs some messages and requires its own instance of `Lo
 #### Loggable derivation
 
 As one can see in [Representation part](#representation) for any of those logged values (like `$request`)  `Loggable`
-instance should be in scope. In the example they are generated by macro, we just need to import it:
+instance should be in scope. In the tofu.example they are generated by macro, we just need to import it:
 
 ```scala
  import tofu.logging.derivation.loggable._
@@ -522,7 +522,7 @@ wrapped in log4cats `Logger`. Moveover, it will log context values that are pass
 as `Map[String, String]`.  
 This means that not only you will not lose your context, but additional values that are passed to `Logger` will also be
 logged.  
-Let's take a look at the example:
+Let's take a look at the tofu.example:
 
 ```scala:reset
 import org.typelevel.log4cats.{Logger, StructuredLogger}
