+++ examples/src/main/scala/tofu/example/doobie/model.scala
@@ -1,4 +1,4 @@
-package tofu.doobie.example
+package tofu.example.doobie
 
 final case class Person(id: Long, name: String, deptId: Long)
 final case class Dept(id: Long, name: String)
