+++ modules/kernel/src/main/scala/tofu/WithRun.scala
@@ -14,7 +14,7 @@ trait WithRun[F[_], G[_], C] extends WithProvide[F, G, C] with WithLocal[F, C] w
 
   /** Allows to convert some context-unaware computation into contextual one.
     *
-    * @example {{{
+    * @tofu.example {{{
     *   trait ProcessHandler[G[_]] {
     *     def mapK[M[_]](fk: G ~> M): ProcessHandler[M] = ???
     *     //...other methods
